Refresh patches.suse/powerpc-pseries-Add-failure-related-checks-for-h_get.patch.
Changing hte return value to long breaks ABI which makes the upstream
version unsuitable for SP6.

suse-commit: 3034e83cd7d3f934dd4c7abcc066b284865a0408

+++ arch/powerpc/platforms/pseries/lpar.c
@@ -1885,7 +1885,7 @@ out:
  * h_get_mpp
  * H_GET_MPP hcall returns info in 7 parms
  */
-int h_get_mpp(struct hvcall_mpp_data *mpp_data)
+long h_get_mpp(struct hvcall_mpp_data *mpp_data)
 {
 	unsigned long retbuf[PLPAR_HCALL9_BUFSIZE] = {0};
 	long rc;
